MercadoArtesanos/requirements.txt
asgiref==3.4.0
cx-Oracle==8.2.1
Django==3.2.4
djangorestframework==3.12.4
Pillow==8.2.0
pytz==2021.1
sqlparse==0.4.1
